Change type of matrix properties from byte to int
According to JOL there is no benefit in object size when using 4 bytes
alignment

+++ src/org/joml/Matrix4dc.java
@@ -140,7 +140,7 @@ public interface Matrix4dc {
      * 
      * @return the properties of the matrix
      */
-    byte properties();
+    int properties();
 
     /**
      * Return the value of the matrix element at column 0 and row 0.
